+++ packages/gamut-illustrations/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.54.9](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-illustrations@0.54.8...@codecademy/gamut-illustrations@0.54.9) (2025-08-20)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
+### [0.54.8](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-illustrations@0.54.7...@codecademy/gamut-illustrations@0.54.8) (2025-08-20)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
 ### [0.54.7](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-illustrations@0.54.6...@codecademy/gamut-illustrations@0.54.7) (2025-07-22)
 
 **Note:** Version bump only for package @codecademy/gamut-illustrations